test(auth): consolidate Gateway regressions and retain method bindings
